Separated convert logic
